Add code skeleton to output sequence of the new codon around mutated site
This roughly sketches places which need to be added to provide codon sequence
and their frequency in the output. I never wrote anything in GO so this is an
incomplete. The pkg/variants/pairwise.go needs more work to figure out where
is the current mutation located in respect to the codon and then slice out the
three nucleotides including current position.